+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [v9.1.0.1006]
 > Release Date: 2026-xx-xx
  - [Changed] Gatered additional NSX Manager Node details for management domain
+ - [Fixed] Security fixes
+ - [Fixed] IP address range manipulation error handling
 
 ## [v9.1.0.1005]
 > Release Date: 2026-06-30